[#1085] Extract find timestamp property from FileStreamProtocol to EventSchemaUtils
