+++ streampipes-extensions-management/pom.xml
@@ -98,6 +98,12 @@
         </dependency>
         <dependency>
             <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-test-utils</artifactId>
+            <version>0.91.0-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-vocabulary</artifactId>
             <version>0.91.0-SNAPSHOT</version>
         </dependency>
