+++ streampipes-extensions-management/src/main/java/org/apache/streampipes/extensions/management/connect/adapter/model/generic/GenericAdapter.java
@@ -67,10 +67,6 @@ public abstract class GenericAdapter<T extends AdapterDescription> extends Adapt
     IProtocol protocolInstance = this.protocol.getInstance(protocolDescription, parser, format);
     this.protocol = protocolInstance;
 
-    //TODO remove
-//    EventSchema eventSchema = adapterDescription.getEventSchema();
-//    this.protocol.setEventSchema(eventSchema);
-
     logger.debug("Start adatper with format: " + format.getId() + " and " + protocol.getId());
 
     protocolInstance.run(adapterPipeline);
