+++ streampipes-model/pom.xml
@@ -40,11 +40,6 @@
         </dependency>
         <dependency>
             <groupId>org.apache.streampipes</groupId>
-            <artifactId>streampipes-test-utils</artifactId>
-            <version>0.91.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-logging</artifactId>
             <version>0.91.0-SNAPSHOT</version>
         </dependency>
