+++ streampipes-model/src/main/java/org/apache/streampipes/model/util/SchemaUtils.java
@@ -19,17 +19,11 @@
 package org.apache.streampipes.model.util;
 
 import org.apache.streampipes.model.schema.EventProperty;
-import org.apache.streampipes.model.schema.EventPropertyNested;
-import org.apache.streampipes.model.schema.EventPropertyPrimitive;
-import org.apache.streampipes.model.schema.EventSchema;
-import org.apache.streampipes.vocabulary.SO;
 
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 public class SchemaUtils {
 
@@ -55,32 +49,4 @@ public class SchemaUtils {
     return properties;
   }
 
-  /**
-   * Returns the timestamp property of an event schema as an {@code Optional}.
-   *
-   * <p> The method checks all properties if they are of type {@code EventPropertyPrimitive} and if their domain
-   * properties contains the uri http://schema.org/DateTime </p>
-   *
-   * @param eventSchema the event schema for which the timestamp property is to be returned
-   * @return an {@code Optional} containing the timestamp property, or an empty {@code Optional} if
-   * no such property was found
-   */
-  public static Optional<EventPropertyPrimitive> getTimestampProperty(EventSchema eventSchema) {
-    return getTimstampProperty(eventSchema.getEventProperties());
-  }
-
-
-  private static Optional<EventPropertyPrimitive> getTimstampProperty(List<EventProperty> eventProperties) {
-    for (EventProperty ep : eventProperties) {
-      if (ep instanceof EventPropertyPrimitive && ep.getDomainProperties().contains(URI.create(SO.DATE_TIME))) {
-        return Optional.of((EventPropertyPrimitive) ep);
-      }
-
-      if (ep instanceof EventPropertyNested) {
-        return getTimstampProperty(((EventPropertyNested) ep).getEventProperties());
-      }
-    }
-
-    return Optional.empty();
-  }
 }
